fix integer sign arithmetic in crypto_verify_*, crypto_stream_salsa20_xor and crypto_onetime_auth methods
